requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
audioread==3.0.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlab==0.8
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
online-judge-api-client==10.10.1
online-judge-tools==11.5.1
online-judge-verify-helper==5.6.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySide2==5.15.2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
shiboken2==5.15.2.1
six==1.16.0
sounddevice==0.4.5
soundfile==0.11.0
soupsieve==2.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1

##### For Whisper AI Transcription #####
SpeechRecognition
openai-whisper
pydub
numba
numpy
torch
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
#sudo apt install ffmpeg

##### For Intent Classification #####
torch                         1.12.0+cpu
nltk

##### For Browser Control #####
pyautogui
pywhatkit
webbrowser

##### For API Calls #####
wolframalpha
wikipedia
pywikihow

##### For webcam control #####
cv2

##### For STT #####
#ovos